Enemies all have their own scripts now.
