+++ Assets/Prefabs/Scuttler.prefab
@@ -181,7 +181,7 @@ GameObject:
   - component: {fileID: 493293538622486411}
   - component: {fileID: 5835613227645742106}
   - component: {fileID: 3044909082382528641}
-  - component: {fileID: 3896103430511743561}
+  - component: {fileID: 5612469087181216065}
   - component: {fileID: 5779878395216687929}
   - component: {fileID: 4907027460689954474}
   - component: {fileID: -6904235409550674936}
@@ -260,7 +260,7 @@ CapsuleCollider:
   m_Height: 2
   m_Direction: 1
   m_Center: {x: 0, y: 1, z: 0}
---- !u!114 &3896103430511743561
+--- !u!114 &5612469087181216065
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -269,7 +269,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 8725256848443516062}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 4a90b18e95dbb224fbb43f2a6e96b4de, type: 3}
+  m_Script: {fileID: 11500000, guid: 155bc444038dc264c8ab99e0df7cd348, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   StartHealth: 38
@@ -279,12 +279,7 @@ MonoBehaviour:
   target: {fileID: 0}
   MovementMode: 1
   TurnRate: 720
-  doubleDash: 0
   WalkSpeed: 30
-  thrust: 30
-  ThrustDelay: 0.75
-  randomFactorRange: 0.1
-  animDelay: 0
   anim: {fileID: 8691365260090520371}
   NextWaypointDistance: 1
   PopupPrefab: {fileID: 5382661844637383376, guid: b9296d04b36c7e2478f7f34e9a8f502c, type: 3}
