+++ Assets/Prefabs/Strider.prefab
@@ -264,7 +264,7 @@ GameObject:
   - component: {fileID: 8783418744222521335}
   - component: {fileID: 7407560020105490601}
   - component: {fileID: 4375632686030564721}
-  - component: {fileID: 3438982514586436998}
+  - component: {fileID: 2614368380816242940}
   - component: {fileID: 1445281374763124501}
   - component: {fileID: 6748171434654086253}
   - component: {fileID: 7530062303969837521}
@@ -343,7 +343,7 @@ CapsuleCollider:
   m_Height: 2.5
   m_Direction: 1
   m_Center: {x: 0, y: 1, z: 0}
---- !u!114 &3438982514586436998
+--- !u!114 &2614368380816242940
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -352,7 +352,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 6652104385415750006}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 4a90b18e95dbb224fbb43f2a6e96b4de, type: 3}
+  m_Script: {fileID: 11500000, guid: 340376a83dd49144b8ff4eafed7cbbda, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   StartHealth: 48
@@ -362,15 +362,15 @@ MonoBehaviour:
   target: {fileID: 0}
   MovementMode: 0
   TurnRate: 720
-  doubleDash: 0
   WalkSpeed: 20
-  thrust: 65
-  ThrustDelay: 0.75
-  randomFactorRange: 0.1
-  animDelay: 0.6
   anim: {fileID: 7119561676194265964}
   NextWaypointDistance: 1
   PopupPrefab: {fileID: 5382661844637383376, guid: b9296d04b36c7e2478f7f34e9a8f502c, type: 3}
+  doubleDash: 0
+  thrust: 45
+  ThrustDelay: 0.75
+  randomFactorRange: 0.1
+  animDelay: 0.6
 --- !u!114 &1445281374763124501
 MonoBehaviour:
   m_ObjectHideFlags: 0
