+++ Assets/Prefabs/Vitriclaw.prefab
@@ -280,7 +280,7 @@ GameObject:
   - component: {fileID: 3196955550826976605}
   - component: {fileID: 6979119148596400993}
   - component: {fileID: 4889046259301199664}
-  - component: {fileID: 6776457321701670076}
+  - component: {fileID: 1337759234801867241}
   - component: {fileID: 5070069568279742733}
   m_Layer: 9
   m_Name: Vitriclaw
@@ -408,7 +408,7 @@ MonoBehaviour:
   raycastOffset: {x: 0, y: 0, z: 0}
   useGraphRaycasting: 1
   quality: 0
---- !u!114 &6776457321701670076
+--- !u!114 &1337759234801867241
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -417,7 +417,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 2276678318342156223}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f5151717131dc5c4286a58c2ecb05d02, type: 3}
+  m_Script: {fileID: 11500000, guid: dce5a79d27b6515478247666e1d26f70, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   StartHealth: 261
@@ -427,20 +427,14 @@ MonoBehaviour:
   target: {fileID: 0}
   MovementMode: 1
   TurnRate: 720
-  doubleDash: 0
   WalkSpeed: 40
-  thrust: 30
-  ThrustDelay: 0.75
-  randomFactorRange: 0.1
-  animDelay: 0.6
   anim: {fileID: 3260272899287883542}
   NextWaypointDistance: 1
-  PopupPrefab: {fileID: 0}
-  attackRange: 7.5
+  PopupPrefab: {fileID: 3144514733173325165, guid: c0bd3be507f85f547b3184d95c4941ad, type: 3}
+  attackRange: 6
   attackTime: 1.67
   speedMultDuringAttack: 0.75
   rotationMultDuringAttack: 0.45
-  canJump: 1
   jumpSpeed: 55
   jumpMaxTime: 0.5
   jumpCooldown: 4.5
