+++ Assets/Prefabs/War Strider.prefab
@@ -297,6 +297,7 @@ GameObject:
   - component: {fileID: 8071052120754423890}
   - component: {fileID: 7341706020362868439}
   - component: {fileID: 6694962938563113733}
+  - component: {fileID: 2057859268995686818}
   - component: {fileID: 5418016976313445348}
   - component: {fileID: 8793151733820749111}
   - component: {fileID: 2363796284379973124}
@@ -376,6 +377,34 @@ CapsuleCollider:
   m_Height: 2.5
   m_Direction: 1
   m_Center: {x: 0, y: 1, z: 0}
+--- !u!114 &2057859268995686818
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6764596011439999448}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 340376a83dd49144b8ff4eafed7cbbda, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  StartHealth: 91
+  StartPH: 9
+  RegenPH: 0.33
+  PlayerDetected: 0
+  target: {fileID: 0}
+  MovementMode: 0
+  TurnRate: 720
+  WalkSpeed: 20
+  anim: {fileID: 6876773263915272632}
+  NextWaypointDistance: 1
+  PopupPrefab: {fileID: 3144514733173325165, guid: c0bd3be507f85f547b3184d95c4941ad, type: 3}
+  doubleDash: 1
+  thrust: 65
+  ThrustDelay: 0.55
+  randomFactorRange: 0.15
+  animDelay: 0.6
 --- !u!114 &5418016976313445348
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -395,12 +424,7 @@ MonoBehaviour:
   target: {fileID: 0}
   MovementMode: 0
   TurnRate: 720
-  doubleDash: 1
   WalkSpeed: 20
-  thrust: 40
-  ThrustDelay: 0.65
-  randomFactorRange: 0.1
-  animDelay: 0
   anim: {fileID: 6876773263915272632}
   NextWaypointDistance: 1
   PopupPrefab: {fileID: 5382661844637383376, guid: b9296d04b36c7e2478f7f34e9a8f502c, type: 3}
