+++ Assets/Scripts/EnemyBehaviorAxehead.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d3a407ca93b6d2b49a7f034268620441
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
