+++ Assets/Scripts/EnemyBehaviorStrider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 340376a83dd49144b8ff4eafed7cbbda
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
