+++ Assets/Scripts/EnemyBehaviorVitriclaw.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dce5a79d27b6515478247666e1d26f70
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
