+++ Assets/Scripts/EnemyBehaviorWalk.cs
@@ -1,0 +1,20 @@
+using Pathfinding;
+using System.Collections;
+using System.Collections.Generic;
+using System.Runtime.CompilerServices;
+using UnityEngine;
+//using UnityEditor.Experimental.GraphView;
+using UnityEngine.UIElements;
+using Unity.VisualScripting.Antlr3.Runtime.Misc;
+using static UnityEngine.GraphicsBuffer;
+
+
+
+public class EnemyBehaviorWalk : EnemyBehavior
+{
+    // Lol the scuttler basically just uses the default enemy controller
+    // Putting this here in case we add anything tho
+
+}
+
+// States
