+++ Assets/Scripts/EnemyBehaviorWalk.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 155bc444038dc264c8ab99e0df7cd348
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
